core: Allow user to config CertificationValidityCheckTimeInHours
Bug-Url: https://bugzilla.redhat.com/1257905

+++ backend/manager/modules/common/src/main/java/org/ovirt/engine/core/common/config/ConfigValues.java
@@ -2149,7 +2149,7 @@ public enum ConfigValues {
     /**
      * The interval in hours of checking the validity of the host's certification
      */
-    @TypeConverterAttribute(Integer.class)
+    @TypeConverterAttribute(Double.class)
     @DefaultValueAttribute("24")
     CertificationValidityCheckTimeInHours,
